# Conflicts:
#	cts/src/ColorPalette.cpp
#	cts/src/SceneGenerator.cpp

+++ src/debug_device/ExtendedQueries.cpp
@@ -406,7 +406,7 @@ static const void * ANARI_LIGHT_directional_visible_info(ANARIDataType paramType
          }
       case 1: // default
          if(paramType == ANARI_BOOL && infoType == ANARI_BOOL) {
-            static const int32_t default_value[1] = {INT32_C(1)};
+            static const int8_t default_value[1] = {INT8_C(1)};
             return default_value;
          } else {
             return nullptr;
@@ -701,7 +701,7 @@ static const void * ANARI_LIGHT_point_visible_info(ANARIDataType paramType, int 
          }
       case 1: // default
          if(paramType == ANARI_BOOL && infoType == ANARI_BOOL) {
-            static const int32_t default_value[1] = {INT32_C(1)};
+            static const int8_t default_value[1] = {INT8_C(1)};
             return default_value;
          } else {
             return nullptr;
@@ -1196,7 +1196,7 @@ static const void * ANARI_LIGHT_ring_visible_info(ANARIDataType paramType, int i
          }
       case 1: // default
          if(paramType == ANARI_BOOL && infoType == ANARI_BOOL) {
-            static const int32_t default_value[1] = {INT32_C(1)};
+            static const int8_t default_value[1] = {INT8_C(1)};
             return default_value;
          } else {
             return nullptr;
@@ -1620,7 +1620,7 @@ static const void * ANARI_LIGHT_quad_visible_info(ANARIDataType paramType, int i
          }
       case 1: // default
          if(paramType == ANARI_BOOL && infoType == ANARI_BOOL) {
-            static const int32_t default_value[1] = {INT32_C(1)};
+            static const int8_t default_value[1] = {INT8_C(1)};
             return default_value;
          } else {
             return nullptr;
@@ -1892,7 +1892,7 @@ static const void * ANARI_LIGHT_hdri_visible_info(ANARIDataType paramType, int i
          }
       case 1: // default
          if(paramType == ANARI_BOOL && infoType == ANARI_BOOL) {
-            static const int32_t default_value[1] = {INT32_C(1)};
+            static const int8_t default_value[1] = {INT8_C(1)};
             return default_value;
          } else {
             return nullptr;
@@ -4172,7 +4172,7 @@ static const void * ANARI_LIGHT_spot_visible_info(ANARIDataType paramType, int i
          }
       case 1: // default
          if(paramType == ANARI_BOOL && infoType == ANARI_BOOL) {
-            static const int32_t default_value[1] = {INT32_C(1)};
+            static const int8_t default_value[1] = {INT8_C(1)};
             return default_value;
          } else {
             return nullptr;
